Updated UI to collapse CC box
Added div around CC box and associated note.  Added checkbox inline with
existing checkbox for "Publicly viewable?"   Then added jquery
statements to clear the textbox and hide the div when the box is NOT
checked and to show the Div when the box is checked.
I repeated the process for the schedule elements.