Implemented the Redeemer look and feel for all Touchpoint Account pages.
The AccountController checks for a setting called UX-LoginPageShell which points to special HTML content that contains the layout of the Redeemer page. Custom styles were written to override the Redeemer/all.css stylesheet that wrecked the look and feel. These styles were placed inline in the special content HTML - with the idea that maybe we could create a separate special content for the CSS file and point to it that way.
A CustomLayout.cshtml view was created so that we didn't have to write the body content twice.