Change contact extra values to support config (#122)
The earlier contact extra values support was based entirely on the three contact dropdowns; however, this was far more complicated, so instead, there is now a config file (i.e. ContactExtraLocationConfig.xml) that allows users to set up specific overrides. If those aren't specified, then the location is based on if the contact is tied to a person or org.

See the unit tests for a brief overview of what is expected and how it should behave. The javascript code is essentially doing the same logic on the front end.