Add new Group Condition for AnyFalse
There are now 4 (instead of 3) group types: AllTrue, AnyTrue, AllFalse, and AnyFalse

AllTrue is displayed to the user as Match All of the conditions below,
AnyTrue is displayed to the user as Match Any of the conditions below
AllFalse is displayed to the user as Match None of the conditions below
AnyFalse is displayed to the user as Match Not All of the conditions below

Use Case of AnyFalse group:
If you have two conditions in a AllTrue group to define primary members as MemberStatusId = Member AND PositionInFamily = Primary,
then you can negate that group by using the AnyFalse group type.
This would be saying NOT MemberStatusId = Member OR NOT PositionInFamily = Primary
which is equivalent to NOT ( MemberStatusId = Member AND PositionInFamily = Primary ).
This allows the user to easily get the complement of the members query.
In other works, the AllTrue would give you primary members, and the AnyFalse would give you the rest of the database.