Fix links for View Code and View Sql links for Search Builder.
If you come to Search Builder from a saved search, you always get a copy of the query for you to experiment with.
Then if you want to update the search, you just Save the query and it will overwrite the original.
But the two links for viewing code were always showing you the code and sql for the original saved search and not for the current query being experimented with.