Don't use DbUtil.ContentFromId(id)
Use CurrentDatabase.Contents.SingleOrDefault(c => c.Id == id) instead
Later, it calls CurrentDatabase.SubmitChanges() which is using a different DataContext.
This code is like this in master and seems to work there,
so there must be another side effect elsewhere that is causing the error in develop.
Anyway, this change fixes it.