More Debugging support for PyScript and PyScriptForm
Adds ability to use a URL like `/PyScriptForm/c!dev-columbarium-ColumbariumRecord.py/person/2`
where the `c!` is translated as `c:/`, the `dev-columbarium-` part is translated as `dev/columbarium/`.
This is is interpreted in debug mode only.
The effect is that you can edit and run from a file on your filesystem rather than having to edit from Special Content.
And it reads this file and runs it in debug mode so that you can set breakpoints in the Python script.
Furthermore, it saves the file contents to a Special Content record based on the extension type.
The name for the Special Content record is `ColumbariumRecord` in this example and is saved under the Python Scripts tab.

A similar technique is used with `model.Content` in debug mode that allows you to read from a file on your computer and have it update the Special Content each time it is read. In this case, you can use regular Path syntax like `model.Content('c:/dev/columbarium/Data/ColumbariumData.json')`. This will get saved to a Special Content record called `ColumbariumData` under the Text tab.

Refactored ScriptController and ScriptModel (now PythonScriptModel and SqlScriptModel) to support the above and removed dependancies on DbUtil.Db.

This also extended to a number of Shared Views that were calling into some static methods in ScriptModel that used DbUtil.Db.
So I also refactored out DbUtil.Db from these Shared Views that refered to DbUtil.Db by adding the following line in RequestManager.cs:
`ViewBag.CurrentDatabase = requestManager.CurrentDatabase;`
This gives all of these Shared Views access to CurrentDatabase as instantiated by the RequestManager constructor.