Used Vue CLI to create a new Vue.js application and merge it into the CmsWeb project. Everything is working. Most of the changes are in the src folder. Added a bunch of packages to make this possible. Added Babel and EsLint. Updated gulp-less from 3.0.1 to 4.0.1. Also installed Axios to access api data from the carious Vue components. Added the vue.config.js file.
Vue CLI will compile, bundle and minify the js and css code for the Vue single file components and put that minified code into the Content/dist folder. This is important. Still lots of work to be done on giving pages but this is a big change in the application and all the plumbing needed to be set up. I finished converting my work on giving pages over from a simple vue instance in the razor page to various Vue components in the src folder in single file components. Lot of work but this sets us up nicely for producing better, cleaner, more cohesive and maintainable code in the future. This also lets us fully take advantage of Vue.js!

Also, we need to do something about including these Vue components in the layout page. I've just put the vue instance that references the single page components in the Layout page for now. Will need to find a solution to this.