Add orderby clause.
This should eliminate almost all problems with duplicate cells since it is unlikely that two Primary Adults (husband and wife) will be using the same Cell phone. The problem seems to always be when children's records are populated with the parents cell phone when adding a new person during check-in. Should be corrected by the designated New People Manager at the church.